reorganisation du code
