+++ received-emails/201219020400758.eml
@@ -1,0 +1,19 @@
+        Fri, 20 Dec 2019 14:04:00 +0100 (CET)
+From: mairie@happytown.com
+To: marie.carin@example.fr
+Message-ID: <1198237193.0.1576847040642.JavaMail.amira.halaoua@CPX-Q7VHWUAZ3OL>
+Subject: Happy Birthday in HappyTown!
+MIME-Version: 1.0
+Content-Type: text/plain; charset=windows-1252
+Content-Transfer-Encoding: quoted-printable
+
+Bonjour Marie Carin,
+
+F=E9licitations, pour f=EAter votre premier anniversaire dans notre merveil=
+leuse ville HappyTown, la mairie a le plaisir de vous offrir un cadeau de b=
+ienvenue.
+
+Votre cadeau est : Le mug qui touille tout seul (Montant : 19.90=80 - R=E9f=
+=E9rence : 6a64d9e7)
+
+L'=E9quipe HappyTown
