+++ received-emails/201219020400867.eml
@@ -1,0 +1,19 @@
+        Fri, 20 Dec 2019 14:04:00 +0100 (CET)
+From: mairie@happytown.com
+To: camille.moulin@example.fr
+Message-ID: <1696544367.1.1576847040815.JavaMail.amira.halaoua@CPX-Q7VHWUAZ3OL>
+Subject: Happy Birthday in HappyTown!
+MIME-Version: 1.0
+Content-Type: text/plain; charset=windows-1252
+Content-Transfer-Encoding: quoted-printable
+
+Bonjour Camille Moulin,
+
+F=E9licitations, pour f=EAter votre premier anniversaire dans notre merveil=
+leuse ville HappyTown, la mairie a le plaisir de vous offrir un cadeau de b=
+ienvenue.
+
+Votre cadeau est : Freddie la luciole (Montant : 15.99=80 - R=E9f=E9rence :=
+ fc02d2df)
+
+L'=E9quipe HappyTown
