+++ received-emails/201219020400913.eml
@@ -1,0 +1,18 @@
+        Fri, 20 Dec 2019 14:04:00 +0100 (CET)
+From: mairie@happytown.com
+To: mairie+service-cadeau@happytown.com
+Message-ID: <2028949540.2.1576847040901.JavaMail.amira.halaoua@CPX-Q7VHWUAZ3OL>
+Subject: 20/12/2019 - Synthese des cadeaux pour envoi
+MIME-Version: 1.0
+Content-Type: text/plain; charset=windows-1252
+Content-Transfer-Encoding: quoted-printable
+
+Bonjour,
+
+Voici la liste r=E9capitulative des cadeaux du jour : =20
+* Marie Carin : Le mug qui touille tout seul (Montant : 19.90=80 - R=E9f=E9=
+rence : 6a64d9e7)=20
+* Camille Moulin : Freddie la luciole (Montant : 15.99=80 - R=E9f=E9rence :=
+ fc02d2df)
+
+Merci!
