+++ received-emails/201219020800190.eml
@@ -1,0 +1,19 @@
+        Fri, 20 Dec 2019 14:08:00 +0100 (CET)
+From: mairie@happytown.com
+To: marie.carin@example.fr
+Message-ID: <835248913.18.1576847280081.JavaMail.amira.halaoua@CPX-Q7VHWUAZ3OL>
+Subject: Happy Birthday in HappyTown!
+MIME-Version: 1.0
+Content-Type: text/plain; charset=windows-1252
+Content-Transfer-Encoding: quoted-printable
+
+Bonjour Marie Carin,
+
+F=E9licitations, pour f=EAter votre premier anniversaire dans notre merveil=
+leuse ville HappyTown, la mairie a le plaisir de vous offrir un cadeau de b=
+ienvenue.
+
+Votre cadeau est : Porte stylo et bloc notes (Montant : 12.60=80 - R=E9f=E9=
+rence : 37c88b3c)
+
+L'=E9quipe HappyTown
