+++ received-emails/201219020800255.eml
@@ -1,0 +1,19 @@
+        Fri, 20 Dec 2019 14:08:00 +0100 (CET)
+From: mairie@happytown.com
+To: camille.moulin@example.fr
+Message-ID: <715446388.19.1576847280227.JavaMail.amira.halaoua@CPX-Q7VHWUAZ3OL>
+Subject: Happy Birthday in HappyTown!
+MIME-Version: 1.0
+Content-Type: text/plain; charset=windows-1252
+Content-Transfer-Encoding: quoted-printable
+
+Bonjour Camille Moulin,
+
+F=E9licitations, pour f=EAter votre premier anniversaire dans notre merveil=
+leuse ville HappyTown, la mairie a le plaisir de vous offrir un cadeau de b=
+ienvenue.
+
+Votre cadeau est : Peluche oggy et les cafards (Montant : 11.99=80 - R=E9f=
+=E9rence : 70d73d02)
+
+L'=E9quipe HappyTown
