+++ received-emails/201219020800289.eml
@@ -1,0 +1,18 @@
+        Fri, 20 Dec 2019 14:08:00 +0100 (CET)
+From: mairie@happytown.com
+To: mairie+service-cadeau@happytown.com
+Message-ID: <57202791.20.1576847280269.JavaMail.amira.halaoua@CPX-Q7VHWUAZ3OL>
+Subject: 20/12/2019 - Synthese des cadeaux pour envoi
+MIME-Version: 1.0
+Content-Type: text/plain; charset=windows-1252
+Content-Transfer-Encoding: quoted-printable
+
+Bonjour,
+
+Voici la liste r=E9capitulative des cadeaux du jour : =20
+* Marie Carin : Porte stylo et bloc notes (Montant : 12.60=80 - R=E9f=E9ren=
+ce : 37c88b3c)=20
+* Camille Moulin : Peluche oggy et les cafards (Montant : 11.99=80 - R=E9f=
+=E9rence : 70d73d02)
+
+Merci!
