+++ received-emails/201219021404054.eml
@@ -1,0 +1,19 @@
+        Fri, 20 Dec 2019 14:14:04 +0100 (CET)
+From: mairie@happytown.com
+To: camille.moulin@example.fr
+Message-ID: <289535655.1.1576847643667.JavaMail.amira.halaoua@CPX-Q7VHWUAZ3OL>
+Subject: Happy Birthday in HappyTown!
+MIME-Version: 1.0
+Content-Type: text/plain; charset=windows-1252
+Content-Transfer-Encoding: quoted-printable
+
+Bonjour Camille Moulin,
+
+F=E9licitations, pour f=EAter votre premier anniversaire dans notre merveil=
+leuse ville HappyTown, la mairie a le plaisir de vous offrir un cadeau de b=
+ienvenue.
+
+Votre cadeau est : La pieuvre d'activit=C3=A9s (Montant : 19.99=80 - R=E9f=
+=E9rence : 66418d5b)
+
+L'=E9quipe HappyTown
