+++ received-emails/201219021404532.eml
@@ -1,0 +1,18 @@
+        Fri, 20 Dec 2019 14:14:04 +0100 (CET)
+From: mairie@happytown.com
+To: mairie+service-cadeau@happytown.com
+Message-ID: <1951618163.2.1576847644085.JavaMail.amira.halaoua@CPX-Q7VHWUAZ3OL>
+Subject: 20/12/2019 - Synthese des cadeaux pour envoi
+MIME-Version: 1.0
+Content-Type: text/plain; charset=windows-1252
+Content-Transfer-Encoding: quoted-printable
+
+Bonjour,
+
+Voici la liste r=E9capitulative des cadeaux du jour : =20
+* Marie Carin : Le mug qui touille tout seul (Montant : 19.90=80 - R=E9f=E9=
+rence : 6a64d9e7)=20
+* Camille Moulin : La pieuvre d'activit=C3=A9s (Montant : 19.99=80 - R=E9f=
+=E9rence : 66418d5b)
+
+Merci!
