+++ src/main/java/com/happytown/core/entities/Cadeau.java
@@ -1,4 +1,4 @@
-package com.happytown.domain;
+package com.happytown.core.entities;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
