+++ src/test/java/com/happytown/service/HabitantServiceTest.java
@@ -1,6 +1,6 @@
 package com.happytown.service;
 
-import com.happytown.domain.Habitant;
+import com.happytown.core.entities.Habitant;
 import com.happytown.fixtures.HabitantFixture;
 import com.happytown.repository.HabitantRepository;
 import org.assertj.core.util.Lists;
